+++ host-configs/rzadams-toss_4_x86_64_ib_cray-llvm-amdgpu@6.3.1_hip.cmake
@@ -4,28 +4,28 @@
 # CMake executable path: /usr/tce/packages/cmake/cmake-3.29.2/bin/cmake
 #------------------------------------------------------------------------------
 
-set(CMAKE_PREFIX_PATH "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/blt-develop-glgaihoun6euvmdtqtlyoceslp5pdqrr;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/c2c-1.8.0-k6bzhviyoy4b4qiwcry5ngq2igehqpxz;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/caliper-git.9ab0e64d2e328d1fdff2ffebb53f30a899d6a598_master-dzlt2f4kgncqnzaehb2exy27b7nijpb4;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/conduit-0.9.5-ohicadtnuc7ou25j2wjeubibgtn3x27w;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/lua-5.4.6-waiv2xo2eqqwohqre65iu6oqzqcshbsd;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/mfem-4.9.0-xlgphhu27dvyu6t6snijepipaan6x5ee;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/adiak-0.4.0-tfjwrwri6ldtkqz5vvmaeay3tv3kju6r;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/libunwind-1.8.3-hupyutxzo74d26y5lmfsxihgfeaurjoq;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/hdf5-1.8.23-r6op3xgaaudlpritbebr2cjx5lw67q5f;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/parmetis-4.0.3-mcmehxox2f37vwp27zjzkn3i74qqnpfz;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/hypre-2.27.0-cc7jeieyge55h5axy5evcdi654jeyl5j;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/raja-git.3b8b59a1e9be2e1066c0d77372b3bf5956e6d6e2_develop-r33uioxc7aidenrvcb35hydewejgfr5e;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/umpire-git.0372fbd6e1f17d7e6dd72693f8b857f3ec7559e9_develop-5jmhwrbu2wqc6qswsjen75xoz4yymyna;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/zlib-1.3.1-xnk7wzwuyxloybgp6m2yhjyf4rhgqet3;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/metis-5.1.0-sa27v5szmjpsatve2ugwku5ok3s7mpon;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/camp-git.a8caefa9f4c811b1a114b4ed2c9b681d40f12325_main-7gcb7xgpawoj2eipuqary4mjrrsyznda;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/fmt-11.0.2-tm7xaljndccnw5eif5jy4fsftqm43ffl;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/none-none/compiler-wrapper-1.0-gsjte4xjarvat5o7b7sdeufsa3qr7mjh;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib_cray/2025_09_25_12_05_48/view/python-3.13.5;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib_cray/2025_09_25_12_05_48/view/python-3.13.5;/collab/usr/gapps/shroud/public/toss_4_x86_64_ib_cray/shroud-0.14.0;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib_cray/2025_09_25_12_05_48/view/python-3.13.5;/usr/tce/packages/cmake/cmake-3.29.2;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib_cray/2025_09_25_12_05_48/gcc-13.3.1/cppcheck-2.18.0-prdufaymm4rf6ugi5f2miqbjclhekg6w;/usr/tce/packages/cray-mpich-tce/cray-mpich-8.1.29-rocmcc-6.3.1;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib_cray/2025_09_25_12_05_48/view/doxygen-1.13.2;/opt/rocm-6.3.1;/opt/rocm-6.3.1;/usr/tce/packages/rocmcc/rocmcc-6.4.2-magic/llvm;/opt/rocm-6.3.1;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib_cray/2025_09_25_12_05_48/view/python-3.13.5;/opt/rocm-6.3.1;/opt/rocm-6.3.1;/opt/rocm-6.3.1" CACHE STRING "")
+set(CMAKE_PREFIX_PATH "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/adiak-0.4.0-c4srnrz5apjetx74dt6xjah63o3a7xcx;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/blt-0.7.1-kzmxfjvxm4drr2qhspckvwee6mrlc3e5;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/c2c-1.8.0-lm64gc55hpxilmtsw7geirys3q4wacu5;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/conduit-0.9.5-wswspowegvp5byl5inc2cikljzkcqg64;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/lua-5.4.8-j7ngytniswhoa536kqfzarcraarkicbm;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/mfem-4.9.0-wetr3wqseqqobjb7be3uzrecpfewq4mr;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/py-nanobind-2.7.0-q2kwozph3gwgfyjnx43jp2gkbaghwxoj;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/none-none/py-pytest-9.0.0-jfdwo62b3p36blpm54gvocrcj6afccv5;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/hdf5-1.8.23-rb4cjthx4wz4wnpaizwmm2jycijdthqm;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/parmetis-4.0.3-ng66tt5fmjyfoiuuxyl4rxqv4ompawmd;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/py-numpy-2.4.2-zcxbrxijjc3miotq6s3kq446afrgycuu;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/hypre-2.27.0-6m2jwyzqbksxgmdznvh54ytysxyme7jy;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/raja-git.3b8b59a1e9be2e1066c0d77372b3bf5956e6d6e2_develop-okzhosj73rts33dgatxh4daotairh5jn;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/umpire-2025.12.0-of6o5lgmd6jgjqfwat2sbyngieagkeco;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/zlib-1.3.1-5w5muvpvct6uruxrddscblsiyizl3uoc;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/metis-5.1.0-4punmdkye6rovr3u7ph7bgj66qim77du;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/camp-git.a8caefa9f4c811b1a114b4ed2c9b681d40f12325_main-74scmbkik6ye7bunoclw46d6zrcfwgha;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/fmt-11.0.2-opsfnufdiedpfpob5wgbvhcrhyhzfkaq;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/none-none/compiler-wrapper-1.0-pypgv2tridcfdliq3cjhfecu23umxcjd;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib_cray/2026_02_17_15_20_25/view/python-3.13.11;/collab/usr/gapps/shroud/public/toss_4_x86_64_ib_cray/shroud-0.14.0;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib_cray/2026_02_17_15_20_25/view/python-3.13.11;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib_cray/2026_02_17_15_20_25/view/python-3.13.11;/usr/tce/packages/cmake/cmake-3.29.2;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib_cray/2026_02_17_15_20_25/gcc-13.3.1/cppcheck-2.18.0-n6kdcwtwlrc3u3t47t7gokpyd4h6mc27;/usr/tce/packages/cray-mpich-tce/cray-mpich-8.1.29-rocmcc-6.3.1;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib_cray/2026_02_17_15_20_25/view/doxygen-1.15.0;/opt/rocm-6.3.1;/opt/rocm-6.3.1;/usr/tce/packages/rocmcc/rocmcc-6.4.3-magic/llvm;/opt/rocm-6.3.1;/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib_cray/2026_02_17_15_20_25/view/python-3.13.11;/opt/rocm-6.3.1;/opt/rocm-6.3.1;/opt/rocm-6.3.1" CACHE STRING "")
 
 set(CMAKE_INSTALL_RPATH_USE_LINK_PATH "ON" CACHE STRING "")
 
-set(CMAKE_BUILD_RPATH "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/axom-develop-64ne2wqvozxbclbji4uxr2u2yimbfknp/lib;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/axom-develop-64ne2wqvozxbclbji4uxr2u2yimbfknp/lib64;;/opt/rh/gcc-toolset-12/root/usr/lib/gcc/x86_64-redhat-linux/12" CACHE STRING "")
+set(CMAKE_BUILD_RPATH "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/axom-develop-xoswouqk3ppxzuxtgegijd3kdoqlkvfe/lib;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/axom-develop-xoswouqk3ppxzuxtgegijd3kdoqlkvfe/lib64;;/opt/rh/gcc-toolset-12/root/usr/lib/gcc/x86_64-redhat-linux/12" CACHE STRING "")
 
-set(CMAKE_INSTALL_RPATH "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/axom-develop-64ne2wqvozxbclbji4uxr2u2yimbfknp/lib;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1/axom-develop-64ne2wqvozxbclbji4uxr2u2yimbfknp/lib64;;/opt/rh/gcc-toolset-12/root/usr/lib/gcc/x86_64-redhat-linux/12" CACHE STRING "")
+set(CMAKE_INSTALL_RPATH "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/axom-develop-xoswouqk3ppxzuxtgegijd3kdoqlkvfe/lib;/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1/axom-develop-xoswouqk3ppxzuxtgegijd3kdoqlkvfe/lib64;;/opt/rh/gcc-toolset-12/root/usr/lib/gcc/x86_64-redhat-linux/12" CACHE STRING "")
 
 set(CMAKE_BUILD_TYPE "Release" CACHE STRING "")
 
 #------------------------------------------------------------------------------
 # Compilers
 #------------------------------------------------------------------------------
-# Compiler Spec: llvm-amdgpu@6.3.1/ewmb23zt2h4lqtnvv6lsha7mdyiucmx3
+# Compiler Spec: llvm-amdgpu@6.3.1/ns2pv5bzkpjolxzmhgkckj6glqymgd7e
 #------------------------------------------------------------------------------
 if(DEFINED ENV{SPACK_CC})
 
-  set(CMAKE_C_COMPILER "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/none-none/compiler-wrapper-1.0-gsjte4xjarvat5o7b7sdeufsa3qr7mjh/libexec/spack/rocmcc/amdclang" CACHE PATH "")
+  set(CMAKE_C_COMPILER "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/none-none/compiler-wrapper-1.0-pypgv2tridcfdliq3cjhfecu23umxcjd/libexec/spack/rocmcc/amdclang" CACHE PATH "")
 
-  set(CMAKE_CXX_COMPILER "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/none-none/compiler-wrapper-1.0-gsjte4xjarvat5o7b7sdeufsa3qr7mjh/libexec/spack/rocmcc/amdclang++" CACHE PATH "")
+  set(CMAKE_CXX_COMPILER "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/none-none/compiler-wrapper-1.0-pypgv2tridcfdliq3cjhfecu23umxcjd/libexec/spack/rocmcc/amdclang++" CACHE PATH "")
 
-  set(CMAKE_Fortran_COMPILER "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/none-none/compiler-wrapper-1.0-gsjte4xjarvat5o7b7sdeufsa3qr7mjh/libexec/spack/rocmcc/amdflang" CACHE PATH "")
+  set(CMAKE_Fortran_COMPILER "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/none-none/compiler-wrapper-1.0-pypgv2tridcfdliq3cjhfecu23umxcjd/libexec/spack/rocmcc/amdflang" CACHE PATH "")
 
 else()
 
@@ -96,52 +96,64 @@ set(ENABLE_GTEST_DEATH_TESTS ON CACHE BOOL "")
 # TPLs
 #------------------------------------------------------------------------------
 
-set(TPL_ROOT "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_01_29_13_42_44/llvm-amdgpu-6.3.1" CACHE PATH "")
+set(TPL_ROOT "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/llvm-amdgpu-6.3.1" CACHE PATH "")
 
-set(CONDUIT_DIR "${TPL_ROOT}/conduit-0.9.5-ohicadtnuc7ou25j2wjeubibgtn3x27w" CACHE PATH "")
+set(CONDUIT_DIR "${TPL_ROOT}/conduit-0.9.5-wswspowegvp5byl5inc2cikljzkcqg64" CACHE PATH "")
 
-set(C2C_DIR "${TPL_ROOT}/c2c-1.8.0-k6bzhviyoy4b4qiwcry5ngq2igehqpxz" CACHE PATH "")
+set(C2C_DIR "${TPL_ROOT}/c2c-1.8.0-lm64gc55hpxilmtsw7geirys3q4wacu5" CACHE PATH "")
 
-set(MFEM_DIR "${TPL_ROOT}/mfem-4.9.0-xlgphhu27dvyu6t6snijepipaan6x5ee" CACHE PATH "")
+set(MFEM_DIR "${TPL_ROOT}/mfem-4.9.0-wetr3wqseqqobjb7be3uzrecpfewq4mr" CACHE PATH "")
 
-set(HDF5_DIR "${TPL_ROOT}/hdf5-1.8.23-r6op3xgaaudlpritbebr2cjx5lw67q5f" CACHE PATH "")
+set(HDF5_DIR "${TPL_ROOT}/hdf5-1.8.23-rb4cjthx4wz4wnpaizwmm2jycijdthqm" CACHE PATH "")
 
-set(LUA_DIR "${TPL_ROOT}/lua-5.4.6-waiv2xo2eqqwohqre65iu6oqzqcshbsd" CACHE PATH "")
+set(LUA_DIR "${TPL_ROOT}/lua-5.4.8-j7ngytniswhoa536kqfzarcraarkicbm" CACHE PATH "")
 
-set(RAJA_DIR "${TPL_ROOT}/raja-git.3b8b59a1e9be2e1066c0d77372b3bf5956e6d6e2_develop-r33uioxc7aidenrvcb35hydewejgfr5e" CACHE PATH "")
+set(RAJA_DIR "${TPL_ROOT}/raja-git.3b8b59a1e9be2e1066c0d77372b3bf5956e6d6e2_develop-okzhosj73rts33dgatxh4daotairh5jn" CACHE PATH "")
 
-set(UMPIRE_DIR "${TPL_ROOT}/umpire-git.0372fbd6e1f17d7e6dd72693f8b857f3ec7559e9_develop-5jmhwrbu2wqc6qswsjen75xoz4yymyna" CACHE PATH "")
+set(UMPIRE_DIR "${TPL_ROOT}/umpire-2025.12.0-of6o5lgmd6jgjqfwat2sbyngieagkeco" CACHE PATH "")
 
 # OPENCASCADE not built
 
-set(ADIAK_DIR "${TPL_ROOT}/adiak-0.4.0-tfjwrwri6ldtkqz5vvmaeay3tv3kju6r" CACHE PATH "")
+set(ADIAK_DIR "${TPL_ROOT}/adiak-0.4.0-c4srnrz5apjetx74dt6xjah63o3a7xcx" CACHE PATH "")
 
-set(CALIPER_DIR "${TPL_ROOT}/caliper-git.9ab0e64d2e328d1fdff2ffebb53f30a899d6a598_master-dzlt2f4kgncqnzaehb2exy27b7nijpb4" CACHE PATH "")
+# CALIPER not built
 
-set(CAMP_DIR "${TPL_ROOT}/camp-git.a8caefa9f4c811b1a114b4ed2c9b681d40f12325_main-7gcb7xgpawoj2eipuqary4mjrrsyznda" CACHE PATH "")
+set(CAMP_DIR "${TPL_ROOT}/camp-git.a8caefa9f4c811b1a114b4ed2c9b681d40f12325_main-74scmbkik6ye7bunoclw46d6zrcfwgha" CACHE PATH "")
 
 # scr not built
 
 #------------------------------------------------------------------------------
-# Devtools
+# Devtools & Python
 #------------------------------------------------------------------------------
 
-set(DEVTOOLS_ROOT "/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib_cray/2025_09_25_12_05_48" CACHE PATH "")
+set(DEVTOOLS_ROOT "/collab/usr/gapps/axom/devtools/toss_4_x86_64_ib_cray/2026_02_17_15_20_25" CACHE PATH "")
 
-set(CLANGFORMAT_EXECUTABLE "/usr/tce/packages/rocmcc/rocmcc-6.4.2-magic/llvm/bin/clang-format" CACHE PATH "")
+set(CLANGFORMAT_EXECUTABLE "/usr/tce/packages/rocmcc/rocmcc-6.4.3-magic/llvm/bin/clang-format" CACHE PATH "")
 
-set(Python_EXECUTABLE "${DEVTOOLS_ROOT}/._view/7qllb33pi46obvxaz6lidqsy2qlr3wkt/python-3.13.5/bin/python3" CACHE PATH "")
+set(Python_EXECUTABLE "${DEVTOOLS_ROOT}/._view/ywmag65dnysd7p4dhlmcoaaqqxfzgz5a/python-3.13.11/bin/python3" CACHE PATH "")
 
-set(JSONSCHEMA_EXECUTABLE "${DEVTOOLS_ROOT}/._view/7qllb33pi46obvxaz6lidqsy2qlr3wkt/python-3.13.5/bin/jsonschema" CACHE PATH "")
+set(JSONSCHEMA_EXECUTABLE "${DEVTOOLS_ROOT}/._view/ywmag65dnysd7p4dhlmcoaaqqxfzgz5a/python-3.13.11/bin/jsonschema" CACHE PATH "")
 
 set(ENABLE_DOCS ON CACHE BOOL "")
 
-set(SPHINX_EXECUTABLE "${DEVTOOLS_ROOT}/._view/7qllb33pi46obvxaz6lidqsy2qlr3wkt/python-3.13.5/bin/sphinx-build" CACHE PATH "")
+set(SPHINX_EXECUTABLE "${DEVTOOLS_ROOT}/._view/ywmag65dnysd7p4dhlmcoaaqqxfzgz5a/python-3.13.11/bin/sphinx-build" CACHE PATH "")
+
+set(YAPF_EXECUTABLE "${DEVTOOLS_ROOT}/._view/ywmag65dnysd7p4dhlmcoaaqqxfzgz5a/python-3.13.11/bin/yapf" CACHE PATH "")
 
 set(SHROUD_EXECUTABLE "/collab/usr/gapps/shroud/public/toss_4_x86_64_ib_cray/shroud-0.14.0/bin/shroud" CACHE PATH "")
 
-set(CPPCHECK_EXECUTABLE "${DEVTOOLS_ROOT}/gcc-13.3.1/cppcheck-2.18.0-prdufaymm4rf6ugi5f2miqbjclhekg6w/bin/cppcheck" CACHE PATH "")
+set(CPPCHECK_EXECUTABLE "${DEVTOOLS_ROOT}/gcc-13.3.1/cppcheck-2.18.0-n6kdcwtwlrc3u3t47t7gokpyd4h6mc27/bin/cppcheck" CACHE PATH "")
 
-set(DOXYGEN_EXECUTABLE "${DEVTOOLS_ROOT}/._view/7qllb33pi46obvxaz6lidqsy2qlr3wkt/doxygen-1.13.2/bin/doxygen" CACHE PATH "")
+set(DOXYGEN_EXECUTABLE "${DEVTOOLS_ROOT}/._view/ywmag65dnysd7p4dhlmcoaaqqxfzgz5a/doxygen-1.15.0/bin/doxygen" CACHE PATH "")
+
+set(PY_NANOBIND_DIR "${TPL_ROOT}/py-nanobind-2.7.0-q2kwozph3gwgfyjnx43jp2gkbaghwxoj/lib/python3.13/site-packages" CACHE PATH "")
+
+set(PY_PYTEST_DIR "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/none-none/py-pytest-9.0.0-jfdwo62b3p36blpm54gvocrcj6afccv5/lib/python3.13/site-packages" CACHE PATH "")
+
+set(PY_NUMPY_DIR "${TPL_ROOT}/py-numpy-2.4.2-zcxbrxijjc3miotq6s3kq446afrgycuu/lib/python3.13/site-packages" CACHE PATH "")
+
+set(PY_PLUGGY_DIR "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/none-none/py-pluggy-1.6.0-n4ojmtqyawxz6gipk64pup3ahwz3xqpu/lib/python3.13/site-packages" CACHE PATH "")
+
+set(PY_INICONFIG_DIR "/usr/WS1/axom/libs/toss_4_x86_64_ib_cray/2026_03_10_11_12_37/none-none/py-iniconfig-2.1.0-lf3yvch65se22jr7b5unfoxtse4ugkrk/lib/python3.13/site-packages" CACHE PATH "")